requirements.txt
pydantic
pyyaml
toml
numpy